+++ packages/db-orchestrator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @evidence-dev/db-orchestrator
 
+## 1.0.0-next.5
+
+### Patch Changes
+
+- 644963c: Add deployment instructions and a check on the status of the user's git repo to the dev mode settings panel
+- Updated dependencies [644963c]
+  - @evidence-dev/bigquery@1.0.0-next.1
+  - @evidence-dev/snowflake@0.0.8-next.1
+  - @evidence-dev/sqlite@1.0.0-next.5
+  - @evidence-dev/telemetry@1.0.0-next.1
+
 ## 1.0.0-next.4
 
 ### Patch Changes