+++ packages/preprocess/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @evidence-dev/preprocess
 
+## 1.0.0-next.2
+
+### Patch Changes
+
+- 644963c: Add deployment instructions and a check on the status of the user's git repo to the dev mode settings panel
+
 ## 1.0.0-next.1
 
 ### Patch Changes