+++ sites/test-env/CHANGELOG.md
@@ -1,5 +1,26 @@
 # evidence-test-environment
 
+## 1.0.0-next.10
+
+### Patch Changes
+
+- @evidence-dev/evidence@1.0.0-next.10
+
+## 1.0.0-next.9
+
+### Patch Changes
+
+- 644963c: Add deployment instructions and a check on the status of the user's git repo to the dev mode settings panel
+- Updated dependencies [644963c]
+  - @evidence-dev/evidence@1.0.0-next.9
+
+## 1.0.0-next.8
+
+### Patch Changes
+
+- Updated dependencies [e45a29a]
+  - @evidence-dev/evidence@1.0.0-next.8
+
 ## 1.0.0-next.7
 
 ### Patch Changes